+++ packages/build-info/src/runtime/bun.ts
@@ -3,5 +3,5 @@ import { LangRuntime } from './runtime.js'
 export class Bun extends LangRuntime {
   id = 'bun'
   name = 'Bun'
-  configFiles = ['bun.lockb', 'bunfig.toml']
+  configFiles = ['bun.lock', 'bun.lockb', 'bunfig.toml']
 }